requirements.txt
# Core Data Analysis
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Jupyter
notebook>=6.5.0
ipykernel>=6.0.0

# Database
mysql-connector-python>=8.0.0

# Airflow (optional — only needed to run the DAG)
# apache-airflow>=2.7.0
# apache-airflow-providers-mysql>=3.0.0
